requirements.txt
pysigma>=0.1.20
pyyaml>=6.0
requests
